[KAN-86] rename: api mypage폴더로 이름 변경
