+++ src/app/layout.tsx
@@ -5,7 +5,7 @@ import type { Metadata } from 'next';
 import { ReactNode } from 'react';
 import Gnb from './components/Gnb/Gnb';
 import Providers from './providers';
-import { getUserData } from './api/actions/user/getUserData';
+import { getUserData } from './api/actions/mypage/getUserData';
 
 export const metadata: Metadata = {
   title: 'Soothe With Me',
